+++ src/Image.cpp
@@ -1,6 +1,6 @@
 #include "Image.h"
 
-Image::Image(SDL_Texture* texture, int x, int y, int w, int h)
+Image::Image(SDL_Texture* texture, int x, int y, uint8_t w, uint8_t h)
   : texture(texture)
   , x(x)
   , y(y)
